Feature: Add RescuableResource module
We started noticing similar patterns throughout the controllers, where, for
example, each `update` action would have to throw an error if the model to
update did not exist.

In order to refactor that, we extracted the logic out so that the controller
would rescue a specific error (in this case `ResourceNotFound`) from any action.

In the end we had three `rescue_from` blocks, and again, after seeing them we
noticed that the blocks could be generalized enough that we didn't have to copy
and paste them into each controller.

Our solution was to create this `RescuableResource` module, which can be enabled
like so:

```ruby
class MyController < ApplicationController
  include Apill::RescuableResource

  rescue_resource 'my_resource_name', from: %w{errors you want rescued},
                                      via:  HumanError::ErrorLookupClass
end
```

The reason we need the `via` option is because each individual engine has its
own configuration for `HumanError` errors and so we can pass in that 'lookup'
class into the `rescue_from` blocks so that each error that ends up being
raised/generated is wrapped with that configuration.