Bugfix: Don't use instance variables for storing the headers
Rails, unbeknownst to us, doesn't create a new instance of the constraint on
every request, therefore if we use instance variables for memoization, they
stick around for the next request.  Due to this, once the first request goes
through, whatever the result of the accept header verification _was_, it
_continues_ to be for each subsequent request.

This change removes any usages of instance variables in favor of local variables
inside `matches?`.