Fixed: Indicies always need their for_query method called
In order to do things like call `total_pages` and `each`, the item has to be
some form of enumerable.  For many indexers, this means calling a `search`
method of some kind to return a result.

Because of this, we always want to call `for_query` on any class whose name
contains `Index` so that we can be sure we're going to get a result object that
we can work with.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: Ic62d170167d44cb27e494fb18ed39df6c1c7b7d9