Added: include to Apill::Resource::Model for accessing in Controllers
This will allow us to access the `include` params passed to the api in the
controller through api_resource.include.

The main goal of this is to remove hard coding the included the associations in
our controllers and keep moving towards the correct JsonApi formatting.

Now we will be able to hit the same controller with different `include` options.

New controller code:
    render json:    api_resource.processed,
           meta:    api_resource.meta,
           include: api_resource.include,
           status:  200

Old Controller code:
    render json:            api_resource.processed,
           meta:            api_resource.meta,
           include:         %w{association_1 association_2 association_2}
           status:          200

JsonApi inclusion of related resources:
http://jsonapi.org/format/#fetching-includes