Added: Tokens::RequestAuthorization
This class handles taking a raw token from a request and verifying it.  It does
this by first decrypting it with our private key and then checking the signature
against our private key to ensure that it hasn't been tampered with.

The result is one of three different classes of tokens:

 * NullRequestAuthoriztion
 * InvalidRequestAuthorization
 * RequestAuthorization

All of these then respond to `to_h`, `valid?` and `blank?` which are used by the
request to decide which token to accept (if any).

For the `RequestAuthorization`, we'll simply return the decrypted, decoded token
data that we can pass up the stack to Rails or whatever.

We can assume that if we're using a `RequestAuthorization` that the token has
passed all checks and is valid.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I3d77b8f7c22cedf30fbedf987f302d82ede726d2