Changed: Requests to allow parsing of authorization tokens
Both the Rails and Rack requests are setup to process authorization tokens
whether they're passed in in the header or in the query params.

The main complexity here is which token to accept based on which one (or both)
are passed in and whether one (or both) are invalid or blank.

Check out the specs for more details.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I8eedb14daeac53beaeab511e9f4654b7a021bf1f