Changed: to_h on null/invalid tokens to an array of hashes
When the valid JWT is decrypted it's going to be an array of two hashes.  The
first hash is going to contain the data and the second hash is going to contain
the headers.

In order for downstream stuff to work, the null/invalid tokens have to respond
with the same type of data structure.

--------------------------------------------------------------------------------
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: Ie63cc861162d3607dfcbe4760d52171f6400f674