Added: token, user and resource to Authorizer::Query
These are the main things a general authorizer is going to need to determine
whether a user/token can access a specific resource.

* `token` is the JSON Web Token data that was used for the request. This will
  typically contain the roles that the user has (eg normal, admin, etc).
* `user` is the actual user object which was looked up for the request
* `resource` is the object which the authorization is requesting access to

The `Query` is also set up so that subclasses may add additional parameters to
the initializer simply by passing them in and adding an `attr_accessor` on the
subclass.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I05e4c32d31c2a156acc38517c94fce5ae6c404a3