Added: Default queries to Authorizer::Query
The typical CRUD action queries have been added.  By default none of them are
authorized with the exception of `index`.  `index` is authorized because very
rarely will a user _not_ be allowed to access and `index`. The authorization of
`index` usually comes in the form of _what_ is listed (the scope) rather than
_can_ I list something.  See a later commit regarding the scope.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I68bfaa1ef72a21222f15569a2feee0925c5518d7