Added: Authorizer::Parameters
When dealing with controllers, the logic behind which parameters are permitted
is very much in line with the authorization that the user is privy to.

For example, maybe an admin has the ability to change the `created_at` time on
a resource, but a regular user cannot.

This will be overridden in each authorizer with the proper implementation.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: Ie5854246c969a01a42572f5ed4353498b7f84ac4