Added: token, user and params to Authorizer::Parameter
These are the main things a parameter authorizer is going to need to determine
which parameters a user/token have access to.

* `token` is the JSON Web Token data that was used for the request. This will
  typically contain the roles that the user has (eg normal, admin, etc).
* `user` is the actual user object which was looked up for the request
* `params` are typically the params passed into the controller from the request

The `Parameter` is also set up so that subclasses may add additional parameters
to the initializer simply by passing them in and adding an `attr_accessor` on
the subclass.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: Id2262900be59becbfc404bd6af6e256abf3395b9