Added: AuthorizableResource
`AuthorizableResource` is where all the magic happens. To be honest this class
does a ton and not much all at the same time.  Not much gets executed in this
class, the main thing that happens is that it calls `authorize` for every
action in the controller.

The reason the class is so long is because it attempts to make it such that very
little configuration is necessary in order to use it.  By default it will use
the name of the controller, the action, and the user, to infer a bunch of
information about the authorizer that needs to be used.

_Note:_ Interestingly none of these 'Resource' classes have specs. Not sure what
I was thinking. I'll add this task to Wunderlist and come back and add them
later.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I8ad487a0e9d7ab3a50838f2141b6fb22c76f7c9f