Added: parameter processing to Authorizers::Parameters::Filtering
This allows the user to have a really easy way to add filtering parameters that
are allowed to be passed in by the user via `add_filtering_parameters` or the
singular version `add_filtering_parameter`.

These methods know how to put a given allowed parameter onto the `filter` params
hash.  It also knows to put an allowed value of an array if the parameter is
being passed in as an array.  This allows for the users of the API to pass in
a scalar value or as an `Array` which is necessary in a lot of cases.

```ruby
params = { :foo => 'bar', :baz => [ :qux, :zaz ] }
```

Then the value generated to be passed to `strong_parameters` will be:

```ruby
[
  :foo,
  {
    baz: [],
  },
]
```

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: I72c1f4386ec56917dedcd83349682beddac20a52