Changed: from passing requested_user_id to params
Passing the `requested_user_id` is too fraught with potential issues.  Not only
that, but due to the way the `Parameters::Filtering` works, the filtering itself
will handle if a specific ID for a specific user is passed in.

The only other case that we need to test for is if we want to default to only
the user's items in the collection when they pass nothing as a param.

For example, if, as an admin, I can see everyone's `Post`s, but as a general
user, I want the site to generally behave the same as it does for a general
user.

In specific cases, I want the API call to return _all_ scouts.

--------------------------------------------------------------------------------
Branch:        feature/add-authorization-logic
Signed-off-by: Jeff Felchner <accounts+git@thekompanee.com>

Change-Id: Ia4987db087056d973e93a3de6780690a5a87e70f